Change Create group to a service
